requirements.txt
# Azure Databricks Knowledge Structuring Pipeline
# Install in Databricks cluster or local dev environment

# ============================================
# LangChain Core
# ============================================
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-experimental>=0.0.50
langgraph>=0.0.20

# ============================================
# Azure SDKs
# ============================================
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-storage-file-datalake>=12.14.0
azure-search-documents>=11.4.0
azure-cosmos>=4.5.0
openai>=1.10.0

# ============================================
# Document Processing
# ============================================
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
extract-msg>=0.48.0
libpst>=0.0.1
python-magic>=0.4.27

# ============================================
# NLP & ML
# ============================================
spacy>=3.7.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
tiktoken>=0.5.0
langdetect>=1.0.9

# ============================================
# GraphRAG
# ============================================
gremlinpython>=3.7.0
igraph>=0.11.0
leidenalg>=0.10.0
networkx>=3.2.0

# ============================================
# Evaluation
# ============================================
ragas>=0.1.0
datasets>=2.16.0

# ============================================
# Data Processing
# ============================================
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0
delta-spark>=3.0.0

# ============================================
# Utilities
# ============================================
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tenacity>=8.2.0
